requirements.txt
numpy~=2.2.6
torch~=2.7.1
librosa~=0.11.0
streamlit~=1.46.1
pydub~=0.25.1
python-dateutil~=2.9.0.post0
pandas~=2.3.0
pymongo~=4.13.2
reportlab~=4.4.2
huggingface-hub~=0.33.1
pydantic~=2.11.7
transformers~=4.53.0
accelerate~=1.8.1
sentencepiece
streamlit-audiorec
openai-whisper
audio_recorder_streamlit